requirements.txt
Django>=5.2.9
djangorestframework>=3.15.0
django-storages>=1.14.0
boto3>=1.34.0
django-redis>=5.4.0
argon2-cffi>=23.1.0
#psycopg2-binary>=2.9.10
segno>=1.6.0
#django-csp>=3.7
django-prometheus>=2.3.1
Pillow>=12.0.0